+++ jib-core/src/main/java/com/google/cloud/tools/jib/registry/BlobPusher.java
@@ -46,6 +46,7 @@ class BlobPusher {
   private final RegistryEndpointRequestProperties registryEndpointRequestProperties;
   private final DescriptorDigest blobDigest;
   private final Blob blob;
+  @Nullable private final String sourceRepository;
 
   /** Initializes the BLOB upload. */
   private class Initializer implements RegistryEndpointProvider<URL> {
@@ -84,11 +85,15 @@ class BlobPusher {
 
     @Override
     public URL getApiRoute(String apiRouteBase) throws MalformedURLException {
-      return new URL(
-          apiRouteBase
-              + registryEndpointRequestProperties.getImageName()
-              + "/blobs/uploads/?mount="
-              + blobDigest);
+      StringBuilder url =
+          new StringBuilder(apiRouteBase)
+              .append(registryEndpointRequestProperties.getImageName())
+              .append("/blobs/uploads/");
+      if (sourceRepository != null) {
+        url.append("?mount=").append(blobDigest).append("&from=").append(sourceRepository);
+      }
+
+      return new URL(url.toString());
     }
 
     @Override
@@ -190,10 +195,12 @@ class BlobPusher {
   BlobPusher(
       RegistryEndpointRequestProperties registryEndpointRequestProperties,
       DescriptorDigest blobDigest,
-      Blob blob) {
+      Blob blob,
+      @Nullable String sourceRepository) {
     this.registryEndpointRequestProperties = registryEndpointRequestProperties;
     this.blobDigest = blobDigest;
     this.blob = blob;
+    this.sourceRepository = sourceRepository;
   }
 
   /**