+++ jib-core/src/main/java/com/google/cloud/tools/jib/registry/RegistryEndpointCaller.java
@@ -19,6 +19,7 @@ package com.google.cloud.tools.jib.registry;
 import com.google.api.client.http.GenericUrl;
 import com.google.api.client.http.HttpResponseException;
 import com.google.api.client.http.HttpStatusCodes;
+import com.google.cloud.tools.jib.JibLogger;
 import com.google.cloud.tools.jib.http.Authorization;
 import com.google.cloud.tools.jib.http.Connection;
 import com.google.cloud.tools.jib.http.Request;
@@ -30,6 +31,7 @@ import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.security.GeneralSecurityException;
 import java.util.function.Function;
 import javax.annotation.Nullable;
 import javax.net.ssl.SSLPeerUnverifiedException;
@@ -51,6 +53,11 @@ class RegistryEndpointCaller<T> {
 
   private static final String DEFAULT_PROTOCOL = "https";
 
+  private static boolean isHttpsProtocol(URL url) {
+    return "https".equals(url.getProtocol());
+  }
+
+  private final JibLogger logger;
   private final URL initialRequestUrl;
   private final String userAgent;
   private final RegistryEndpointProvider<T> registryEndpointProvider;
@@ -67,6 +74,7 @@ class RegistryEndpointCaller<T> {
   /**
    * Constructs with parameters for making the request.
    *
+   * @param logger the build logger used for printing messages
    * @param userAgent {@code User-Agent} header to send with the request
    * @param apiRouteBase the endpoint's API root, without the protocol
    * @param registryEndpointProvider the {@link RegistryEndpointProvider} to the endpoint
@@ -76,6 +84,7 @@ class RegistryEndpointCaller<T> {
    * @throws MalformedURLException if the URL generated for the endpoint is malformed
    */
   RegistryEndpointCaller(
+      JibLogger logger,
       String userAgent,
       String apiRouteBase,
       RegistryEndpointProvider<T> registryEndpointProvider,
@@ -84,6 +93,7 @@ class RegistryEndpointCaller<T> {
       boolean allowInsecureRegistries)
       throws MalformedURLException {
     this(
+        logger,
         userAgent,
         apiRouteBase,
         registryEndpointProvider,
@@ -96,6 +106,7 @@ class RegistryEndpointCaller<T> {
 
   @VisibleForTesting
   RegistryEndpointCaller(
+      JibLogger logger,
       String userAgent,
       String apiRouteBase,
       RegistryEndpointProvider<T> registryEndpointProvider,
@@ -105,6 +116,7 @@ class RegistryEndpointCaller<T> {
       Function<URL, Connection> connectionFactory,
       @Nullable Function<URL, Connection> insecureConnectionFactory)
       throws MalformedURLException {
+    this.logger = logger;
     this.initialRequestUrl =
         registryEndpointProvider.getApiRoute(DEFAULT_PROTOCOL + "://" + apiRouteBase);
     this.userAgent = userAgent;
@@ -125,7 +137,52 @@ class RegistryEndpointCaller<T> {
    */
   @Nullable
   T call() throws IOException, RegistryException {
-    return call(initialRequestUrl);
+    return callWithAllowInsecureRegistryHandling(initialRequestUrl);
+  }
+
+  @Nullable
+  private T callWithAllowInsecureRegistryHandling(URL url) throws IOException, RegistryException {
+    if (!isHttpsProtocol(url) && !allowInsecureRegistries) {
+      throw new InsecureRegistryException(url);
+    }
+
+    try {
+      return call(url, connectionFactory);
+
+    } catch (SSLPeerUnverifiedException ex) {
+      return handleUnverifiableServerException(url);
+    }
+  }
+
+  @Nullable
+  private T handleUnverifiableServerException(URL url) throws IOException, RegistryException {
+    if (!allowInsecureRegistries) {
+      throw new InsecureRegistryException(url);
+    }
+
+    try {
+      return call(url, getInsecureConnectionFactory());
+
+    } catch (SSLPeerUnverifiedException | HttpHostConnectException ex) {
+      // Try HTTP as a last resort.
+      GenericUrl httpUrl = new GenericUrl(url);
+      httpUrl.setScheme("http");
+      logger.warn(
+          "Failed to connect to " + url + " over HTTPS. Attempting again with HTTP: " + httpUrl);
+      return call(httpUrl.toURL(), connectionFactory);
+    }
+  }
+
+  private Function<URL, Connection> getInsecureConnectionFactory() throws RegistryException {
+    try {
+      if (insecureConnectionFactory == null) {
+        insecureConnectionFactory = Connection.getInsecureConnectionFactory();
+      }
+      return insecureConnectionFactory;
+
+    } catch (GeneralSecurityException ex) {
+      throw new RegistryException("cannot turn off TLS peer verification", ex);
+    }
   }
 
   /**
@@ -136,15 +193,11 @@ class RegistryEndpointCaller<T> {
    * @throws IOException for most I/O exceptions when making the request
    * @throws RegistryException for known exceptions when interacting with the registry
    */
-  @VisibleForTesting
   @Nullable
-  T call(URL url) throws IOException, RegistryException {
-    boolean isHttpProtocol = "http".equals(url.getProtocol());
-    if (!allowInsecureRegistries && isHttpProtocol) {
-      throw new InsecureRegistryException(url);
-    }
+  private T call(URL url, Function<URL, Connection> connectionFactory)
+      throws IOException, RegistryException {
     // Only sends authorization if using HTTPS or explicitly forcing over HTTP.
-    boolean sendCredentials = !isHttpProtocol || Boolean.getBoolean("sendCredentialsOverHttp");
+    boolean sendCredentials = isHttpsProtocol(url) || Boolean.getBoolean("sendCredentialsOverHttp");
 
     try (Connection connection = connectionFactory.apply(url)) {
       Request.Builder requestBuilder =
@@ -211,25 +264,13 @@ class RegistryEndpointCaller<T> {
             || httpResponseException.getStatusCode() == STATUS_CODE_PERMANENT_REDIRECT) {
           // 'Location' header can be relative or absolute.
           URL redirectLocation = new URL(url, httpResponseException.getHeaders().getLocation());
-          return call(redirectLocation);
+          return callWithAllowInsecureRegistryHandling(redirectLocation);
 
         } else {
           // Unknown
           throw httpResponseException;
         }
       }
-
-    } catch (HttpHostConnectException | SSLPeerUnverifiedException ex) {
-      // Tries to call with HTTP protocol if HTTPS failed to connect.
-      // Note that this will not succeed if 'allowInsecureRegistries' is false.
-      if ("https".equals(url.getProtocol())) {
-        GenericUrl httpUrl = new GenericUrl(url);
-        httpUrl.setScheme("http");
-        return call(httpUrl.toURL());
-      }
-
-      throw ex;
-
     } catch (NoHttpResponseException ex) {
       throw new RegistryNoResponseException(ex);
     }