+++ jib-core/src/test/java/com/google/cloud/tools/jib/registry/RegistryEndpointCallerTest.java
@@ -20,19 +20,16 @@ import com.google.api.client.http.HttpHeaders;
 import com.google.api.client.http.HttpResponse;
 import com.google.api.client.http.HttpResponseException;
 import com.google.api.client.http.HttpStatusCodes;
+import com.google.cloud.tools.jib.JibLogger;
 import com.google.cloud.tools.jib.blob.Blobs;
 import com.google.cloud.tools.jib.http.Authorizations;
 import com.google.cloud.tools.jib.http.BlobHttpContent;
 import com.google.cloud.tools.jib.http.Connection;
 import com.google.cloud.tools.jib.http.MockConnection;
 import com.google.cloud.tools.jib.http.Response;
-import com.google.cloud.tools.jib.json.JsonTemplateMapper;
-import com.google.cloud.tools.jib.registry.json.ErrorEntryTemplate;
-import com.google.cloud.tools.jib.registry.json.ErrorResponseTemplate;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.security.GeneralSecurityException;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Function;
@@ -91,20 +88,36 @@ public class RegistryEndpointCallerTest {
     }
   }
 
+  private static HttpResponse mockHttpResponse(int statusCode, @Nullable HttpHeaders headers)
+      throws IOException {
+    HttpResponse mock = Mockito.mock(HttpResponse.class);
+    Mockito.when(mock.getStatusCode()).thenReturn(statusCode);
+    Mockito.when(mock.parseAsString()).thenReturn("");
+    Mockito.when(mock.getHeaders()).thenReturn(headers != null ? headers : new HttpHeaders());
+    return mock;
+  }
+
+  private static HttpResponse mockRedirectHttpResponse(String redirectLocation) throws IOException {
+    int code307 = HttpStatusCodes.STATUS_CODE_TEMPORARY_REDIRECT;
+    return mockHttpResponse(code307, new HttpHeaders().setLocation(redirectLocation));
+  }
+
+  @Mock private JibLogger mockBuildLogger;
   @Mock private Connection mockConnection;
+  @Mock private Connection mockInsecureConnection;
   @Mock private Response mockResponse;
   @Mock private Function<URL, Connection> mockConnectionFactory;
-  @Mock private HttpResponse mockHttpResponse;
+  @Mock private Function<URL, Connection> mockInsecureConnectionFactory;
 
-  private RegistryEndpointCaller<String> testRegistryEndpointCallerSecure;
+  private RegistryEndpointCaller<String> secureEndpointCaller;
 
   @Before
   public void setUp() throws IOException {
-    testRegistryEndpointCallerSecure = createRegistryEndpointCaller(false);
+    secureEndpointCaller = createRegistryEndpointCaller(false);
 
     Mockito.when(mockConnectionFactory.apply(Mockito.any())).thenReturn(mockConnection);
-    Mockito.when(mockHttpResponse.parseAsString()).thenReturn("");
-    Mockito.when(mockHttpResponse.getHeaders()).thenReturn(new HttpHeaders());
+    Mockito.when(mockInsecureConnectionFactory.apply(Mockito.any()))
+        .thenReturn(mockInsecureConnection);
     Mockito.when(mockResponse.getBody()).thenReturn(Blobs.from("body"));
   }
 
@@ -115,15 +128,91 @@ public class RegistryEndpointCallerTest {
   }
 
   @Test
-  public void testCall_httpsPeerUnverified()
-      throws IOException, RegistryException, GeneralSecurityException {
-    verifyRetriesWithHttp(SSLPeerUnverifiedException.class);
+  public void testCall_secureCallerOnUnverifiableServer() throws IOException, RegistryException {
+    Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)); // unverifiable HTTPS server
+
+    try {
+      secureEndpointCaller.call();
+      Assert.fail("Secure caller should fail if cannot verify server");
+    } catch (InsecureRegistryException ex) {
+      Assert.assertEquals(
+          "Failed to verify the server at https://apiRouteBase/api because only secure connections are allowed.",
+          ex.getMessage());
+    }
   }
 
   @Test
-  public void testCall_retryWithHttp()
-      throws IOException, RegistryException, GeneralSecurityException {
-    verifyRetriesWithHttp(HttpHostConnectException.class);
+  public void testCall_insecureCallerOnUnverifiableServer() throws IOException, RegistryException {
+    Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)); // unverifiable HTTPS server
+    Mockito.when(mockInsecureConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenReturn(mockResponse); // OK with non-verifying connection
+
+    RegistryEndpointCaller<String> insecureCaller = createRegistryEndpointCaller(true);
+    Assert.assertEquals("body", insecureCaller.call());
+
+    ArgumentCaptor<URL> urlCaptor = ArgumentCaptor.forClass(URL.class);
+    Mockito.verify(mockConnectionFactory, Mockito.times(1)).apply(urlCaptor.capture());
+    Assert.assertEquals(new URL("https://apiRouteBase/api"), urlCaptor.getAllValues().get(0));
+
+    Mockito.verify(mockInsecureConnectionFactory, Mockito.times(1)).apply(urlCaptor.capture());
+    Assert.assertEquals(new URL("https://apiRouteBase/api"), urlCaptor.getAllValues().get(1));
+  }
+
+  @Test
+  public void testCall_insecureCallerOnHttpServer() throws IOException, RegistryException {
+    Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)) // server is not HTTPS
+        .thenReturn(mockResponse);
+    Mockito.when(mockInsecureConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)); // server is not HTTPS
+
+    RegistryEndpointCaller<String> insecureEndpointCaller = createRegistryEndpointCaller(true);
+    Assert.assertEquals("body", insecureEndpointCaller.call());
+
+    ArgumentCaptor<URL> urlCaptor = ArgumentCaptor.forClass(URL.class);
+    Mockito.verify(mockConnectionFactory, Mockito.times(2)).apply(urlCaptor.capture());
+    Assert.assertEquals(new URL("https://apiRouteBase/api"), urlCaptor.getAllValues().get(0));
+    Assert.assertEquals(new URL("http://apiRouteBase/api"), urlCaptor.getAllValues().get(1));
+
+    Mockito.verify(mockInsecureConnectionFactory, Mockito.times(1)).apply(urlCaptor.capture());
+    Assert.assertEquals(new URL("https://apiRouteBase/api"), urlCaptor.getAllValues().get(2));
+
+    Mockito.verify(mockBuildLogger)
+        .warn(
+            "Failed to connect to "
+                + urlCaptor.getAllValues().get(0)
+                + " over HTTPS. Attempting again with HTTP: "
+                + urlCaptor.getAllValues().get(1));
+  }
+
+  @Test
+  public void testCall_insecureCallerOnHttpServerByHttpHostConnectException()
+      throws IOException, RegistryException {
+    Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)) // server is not HTTPS
+        .thenReturn(mockResponse);
+    Mockito.when(mockInsecureConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(HttpHostConnectException.class)); // server is not HTTPS
+
+    RegistryEndpointCaller<String> insecureEndpointCaller = createRegistryEndpointCaller(true);
+    Assert.assertEquals("body", insecureEndpointCaller.call());
+
+    ArgumentCaptor<URL> urlCaptor = ArgumentCaptor.forClass(URL.class);
+    Mockito.verify(mockConnectionFactory, Mockito.times(2)).apply(urlCaptor.capture());
+    Assert.assertEquals(new URL("https://apiRouteBase/api"), urlCaptor.getAllValues().get(0));
+    Assert.assertEquals(new URL("http://apiRouteBase/api"), urlCaptor.getAllValues().get(1));
+
+    Mockito.verify(mockInsecureConnectionFactory, Mockito.times(1)).apply(urlCaptor.capture());
+    Assert.assertEquals(new URL("https://apiRouteBase/api"), urlCaptor.getAllValues().get(2));
+
+    Mockito.verify(mockBuildLogger)
+        .warn(
+            "Failed to connect to "
+                + urlCaptor.getAllValues().get(0)
+                + " over HTTPS. Attempting again with HTTP: "
+                + urlCaptor.getAllValues().get(1));
   }
 
   @Test
@@ -134,7 +223,7 @@ public class RegistryEndpointCallerTest {
         .thenThrow(mockNoHttpResponseException);
 
     try {
-      testRegistryEndpointCallerSecure.call();
+      secureEndpointCaller.call();
       Assert.fail("Call should have failed");
 
     } catch (RegistryNoResponseException ex) {
@@ -148,18 +237,21 @@ public class RegistryEndpointCallerTest {
   }
 
   @Test
-  public void testCall_credentialsNotSent() throws IOException, RegistryException {
-    Mockito.when(mockHttpResponse.getStatusCode())
-        .thenReturn(HttpStatusCodes.STATUS_CODE_UNAUTHORIZED);
+  public void testCall_credentialsNotSentOverHttp() throws IOException, RegistryException {
+    HttpResponse redirectResponse = mockRedirectHttpResponse("http://newlocation");
+    HttpResponse unauthroizedResponse =
+        mockHttpResponse(HttpStatusCodes.STATUS_CODE_UNAUTHORIZED, null);
 
-    HttpResponseException httpResponseException = new HttpResponseException(mockHttpResponse);
     Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
-        .thenThrow(httpResponseException);
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)) // server is not HTTPS
+        .thenThrow(new HttpResponseException(redirectResponse)) // redirect to HTTP
+        .thenThrow(new HttpResponseException(unauthroizedResponse)); // final response
+    Mockito.when(mockInsecureConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)); // server is not HTTPS
 
-    RegistryEndpointCaller<String> testRegistryEndpointCallerInsecure =
-        createRegistryEndpointCaller(true);
+    RegistryEndpointCaller<String> insecureEndpointCaller = createRegistryEndpointCaller(true);
     try {
-      testRegistryEndpointCallerInsecure.call(new URL("http://location"));
+      insecureEndpointCaller.call();
       Assert.fail("Call should have failed");
 
     } catch (RegistryCredentialsNotSentException ex) {
@@ -170,15 +262,19 @@ public class RegistryEndpointCallerTest {
   }
 
   @Test
-  public void testCall_sendingCredentialsForced() throws IOException, RegistryException {
-    Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
-        .thenReturn(mockResponse);
+  public void testCall_credentialsForcedOverHttp() throws IOException, RegistryException {
+    HttpResponse redirectResponse = mockRedirectHttpResponse("http://newlocation");
 
-    RegistryEndpointCaller<String> testRegistryEndpointCallerInsecure =
-        createRegistryEndpointCaller(true);
+    Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)) // server is not HTTPS
+        .thenThrow(new HttpResponseException(redirectResponse)) // redirect to HTTP
+        .thenReturn(mockResponse); // final response
+    Mockito.when(mockInsecureConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
+        .thenThrow(Mockito.mock(SSLPeerUnverifiedException.class)); // server is not HTTPS
+
     System.setProperty("sendCredentialsOverHttp", "true");
-    Assert.assertEquals(
-        "body", testRegistryEndpointCallerInsecure.call(new URL("http://location")));
+    RegistryEndpointCaller<String> insecureEndpointCaller = createRegistryEndpointCaller(true);
+    Assert.assertEquals("body", insecureEndpointCaller.call());
   }
 
   @Test
@@ -203,15 +299,15 @@ public class RegistryEndpointCallerTest {
 
   @Test
   public void testCall_unknown() throws IOException, RegistryException {
-    Mockito.when(mockHttpResponse.getStatusCode())
-        .thenReturn(HttpStatusCodes.STATUS_CODE_SERVER_ERROR);
+    HttpResponse mockHttpResponse =
+        mockHttpResponse(HttpStatusCodes.STATUS_CODE_SERVER_ERROR, null);
     HttpResponseException httpResponseException = new HttpResponseException(mockHttpResponse);
 
     Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
         .thenThrow(httpResponseException);
 
     try {
-      testRegistryEndpointCallerSecure.call();
+      secureEndpointCaller.call();
       Assert.fail("Call should have failed");
 
     } catch (HttpResponseException ex) {
@@ -220,38 +316,31 @@ public class RegistryEndpointCallerTest {
   }
 
   @Test
-  public void testCall_temporaryRedirect()
-      throws IOException, RegistryException, GeneralSecurityException {
+  public void testCall_temporaryRedirect() throws IOException, RegistryException {
     verifyRetriesWithNewLocation(HttpStatusCodes.STATUS_CODE_TEMPORARY_REDIRECT);
   }
 
   @Test
-  public void testCall_movedPermanently()
-      throws IOException, RegistryException, GeneralSecurityException {
+  public void testCall_movedPermanently() throws IOException, RegistryException {
     verifyRetriesWithNewLocation(HttpStatusCodes.STATUS_CODE_MOVED_PERMANENTLY);
   }
 
   @Test
-  public void testCall_permanentRedirect()
-      throws IOException, RegistryException, GeneralSecurityException {
+  public void testCall_permanentRedirect() throws IOException, RegistryException {
     verifyRetriesWithNewLocation(RegistryEndpointCaller.STATUS_CODE_PERMANENT_REDIRECT);
   }
 
   @Test
   public void testCall_disallowInsecure() throws IOException, RegistryException {
     // Mocks a response for temporary redirect to a new location.
-    Mockito.when(mockHttpResponse.getStatusCode())
-        .thenReturn(HttpStatusCodes.STATUS_CODE_TEMPORARY_REDIRECT);
-    Mockito.when(mockHttpResponse.getHeaders())
-        .thenReturn(new HttpHeaders().setLocation("http://newlocation"));
+    HttpResponse redirectResponse = mockRedirectHttpResponse("http://newlocation");
 
-    HttpResponseException httpResponseException = new HttpResponseException(mockHttpResponse);
+    HttpResponseException redirectException = new HttpResponseException(redirectResponse);
     Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
-        .thenThrow(httpResponseException)
-        .thenReturn(mockResponse);
+        .thenThrow(redirectException);
 
     try {
-      testRegistryEndpointCallerSecure.call();
+      secureEndpointCaller.call();
       Assert.fail("Call should have failed");
 
     } catch (InsecureRegistryException ex) {
@@ -260,13 +349,12 @@ public class RegistryEndpointCallerTest {
   }
 
   @Test
-  public void testHttpTimeout_propertyNotSet()
-      throws IOException, RegistryException, GeneralSecurityException {
+  public void testHttpTimeout_propertyNotSet() throws IOException, RegistryException {
     MockConnection mockConnection = new MockConnection((httpMethod, request) -> mockResponse);
     Mockito.when(mockConnectionFactory.apply(Mockito.any())).thenReturn(mockConnection);
 
     Assert.assertNull(System.getProperty("jib.httpTimeout"));
-    testRegistryEndpointCallerSecure.call();
+    secureEndpointCaller.call();
 
     // We fall back to the default timeout:
     // https://github.com/GoogleContainerTools/jib/pull/656#discussion_r203562639
@@ -274,25 +362,23 @@ public class RegistryEndpointCallerTest {
   }
 
   @Test
-  public void testHttpTimeout_stringValue()
-      throws IOException, RegistryException, GeneralSecurityException {
+  public void testHttpTimeout_stringValue() throws IOException, RegistryException {
     MockConnection mockConnection = new MockConnection((httpMethod, request) -> mockResponse);
     Mockito.when(mockConnectionFactory.apply(Mockito.any())).thenReturn(mockConnection);
 
     System.setProperty("jib.httpTimeout", "random string");
-    testRegistryEndpointCallerSecure.call();
+    secureEndpointCaller.call();
 
     Assert.assertNull(mockConnection.getRequestedHttpTimeout());
   }
 
   @Test
-  public void testHttpTimeout_negativeValue()
-      throws IOException, RegistryException, GeneralSecurityException {
+  public void testHttpTimeout_negativeValue() throws IOException, RegistryException {
     MockConnection mockConnection = new MockConnection((httpMethod, request) -> mockResponse);
     Mockito.when(mockConnectionFactory.apply(Mockito.any())).thenReturn(mockConnection);
 
     System.setProperty("jib.httpTimeout", "-1");
-    testRegistryEndpointCallerSecure.call();
+    secureEndpointCaller.call();
 
     // We let the negative value pass through:
     // https://github.com/GoogleContainerTools/jib/pull/656#discussion_r203562639
@@ -300,47 +386,27 @@ public class RegistryEndpointCallerTest {
   }
 
   @Test
-  public void testHttpTimeout_0accepted()
-      throws IOException, RegistryException, GeneralSecurityException {
+  public void testHttpTimeout_0accepted() throws IOException, RegistryException {
     System.setProperty("jib.httpTimeout", "0");
 
     MockConnection mockConnection = new MockConnection((httpMethod, request) -> mockResponse);
     Mockito.when(mockConnectionFactory.apply(Mockito.any())).thenReturn(mockConnection);
 
-    testRegistryEndpointCallerSecure.call();
+    secureEndpointCaller.call();
 
     Assert.assertEquals(Integer.valueOf(0), mockConnection.getRequestedHttpTimeout());
   }
 
   @Test
-  public void testHttpTimeout() throws IOException, RegistryException, GeneralSecurityException {
+  public void testHttpTimeout() throws IOException, RegistryException {
     System.setProperty("jib.httpTimeout", "7593");
 
     MockConnection mockConnection = new MockConnection((httpMethod, request) -> mockResponse);
     Mockito.when(mockConnectionFactory.apply(Mockito.any())).thenReturn(mockConnection);
 
-    testRegistryEndpointCallerSecure.call();
+    secureEndpointCaller.call();
 
     Assert.assertEquals(Integer.valueOf(7593), mockConnection.getRequestedHttpTimeout());
-  }
-
-  /** Verifies a request is retried with HTTP protocol if {@code exceptionClass} is thrown. */
-  private void verifyRetriesWithHttp(Class<? extends Throwable> exceptionClass)
-      throws IOException, RegistryException, GeneralSecurityException {
-    // Has mockConnection.send throw first, then succeed.
-    Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
-        .thenThrow(Mockito.mock(exceptionClass))
-        .thenReturn(mockResponse);
-
-    RegistryEndpointCaller<String> testRegistryEndpointCallerInsecure =
-        createRegistryEndpointCaller(true);
-    Assert.assertEquals("body", testRegistryEndpointCallerInsecure.call());
-
-    // Checks that the URL protocol was first HTTPS, then HTTP.
-    ArgumentCaptor<URL> urlArgumentCaptor = ArgumentCaptor.forClass(URL.class);
-    Mockito.verify(mockConnectionFactory, Mockito.times(2)).apply(urlArgumentCaptor.capture());
-    Assert.assertEquals("https", urlArgumentCaptor.getAllValues().get(0).getProtocol());
-    Assert.assertEquals("http", urlArgumentCaptor.getAllValues().get(1).getProtocol());
   }
 
   /**
@@ -349,14 +415,14 @@ public class RegistryEndpointCallerTest {
    */
   private void verifyThrowsRegistryUnauthorizedException(int httpStatusCode)
       throws IOException, RegistryException {
-    Mockito.when(mockHttpResponse.getStatusCode()).thenReturn(httpStatusCode);
+    HttpResponse mockHttpResponse = mockHttpResponse(httpStatusCode, null);
     HttpResponseException httpResponseException = new HttpResponseException(mockHttpResponse);
 
     Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
         .thenThrow(httpResponseException);
 
     try {
-      testRegistryEndpointCallerSecure.call();
+      secureEndpointCaller.call();
       Assert.fail("Call should have failed");
 
     } catch (RegistryUnauthorizedException ex) {
@@ -372,19 +438,16 @@ public class RegistryEndpointCallerTest {
    */
   private void verifyThrowsRegistryErrorException(int httpStatusCode)
       throws IOException, RegistryException {
-    ErrorResponseTemplate errorResponseTemplate =
-        new ErrorResponseTemplate().addError(new ErrorEntryTemplate("code", "message"));
-
-    Mockito.when(mockHttpResponse.getStatusCode()).thenReturn(httpStatusCode);
-    Mockito.when(mockHttpResponse.parseAsString())
-        .thenReturn(Blobs.writeToString(JsonTemplateMapper.toBlob(errorResponseTemplate)));
-    HttpResponseException httpResponseException = new HttpResponseException(mockHttpResponse);
+    HttpResponse errorResponse = mockHttpResponse(httpStatusCode, null);
+    Mockito.when(errorResponse.parseAsString())
+        .thenReturn("{\"errors\":[{\"code\":\"code\",\"message\":\"message\"}]}");
+    HttpResponseException httpResponseException = new HttpResponseException(errorResponse);
 
     Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
         .thenThrow(httpResponseException);
 
     try {
-      testRegistryEndpointCallerSecure.call();
+      secureEndpointCaller.call();
       Assert.fail("Call should have failed");
 
     } catch (RegistryErrorException ex) {
@@ -400,19 +463,18 @@ public class RegistryEndpointCallerTest {
    * Location} header.
    */
   private void verifyRetriesWithNewLocation(int httpStatusCode)
-      throws IOException, RegistryException, GeneralSecurityException {
+      throws IOException, RegistryException {
     // Mocks a response for temporary redirect to a new location.
-    Mockito.when(mockHttpResponse.getStatusCode()).thenReturn(httpStatusCode);
-    Mockito.when(mockHttpResponse.getHeaders())
-        .thenReturn(new HttpHeaders().setLocation("https://newlocation"));
+    HttpResponse redirectResponse =
+        mockHttpResponse(httpStatusCode, new HttpHeaders().setLocation("https://newlocation"));
 
     // Has mockConnection.send throw first, then succeed.
-    HttpResponseException httpResponseException = new HttpResponseException(mockHttpResponse);
+    HttpResponseException redirectException = new HttpResponseException(redirectResponse);
     Mockito.when(mockConnection.send(Mockito.eq("httpMethod"), Mockito.any()))
-        .thenThrow(httpResponseException)
+        .thenThrow(redirectException)
         .thenReturn(mockResponse);
 
-    Assert.assertEquals("body", testRegistryEndpointCallerSecure.call());
+    Assert.assertEquals("body", secureEndpointCaller.call());
 
     // Checks that the URL was changed to the new location.
     ArgumentCaptor<URL> urlArgumentCaptor = ArgumentCaptor.forClass(URL.class);
@@ -425,6 +487,7 @@ public class RegistryEndpointCallerTest {
   private RegistryEndpointCaller<String> createRegistryEndpointCaller(boolean allowInsecure)
       throws MalformedURLException {
     return new RegistryEndpointCaller<>(
+        mockBuildLogger,
         "userAgent",
         "apiRouteBase",
         new TestRegistryEndpointProvider(),
@@ -432,6 +495,6 @@ public class RegistryEndpointCallerTest {
         new RegistryEndpointRequestProperties("serverUrl", "imageName"),
         allowInsecure,
         mockConnectionFactory,
-        null);
+        mockInsecureConnectionFactory);
   }
 }