+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildDockerMojo.java
@@ -60,8 +60,8 @@ public class BuildDockerMojo extends JibPluginConfiguration {
       return;
     }
 
-    MavenBuildLogger mavenBuildLogger = new MavenBuildLogger(getLog());
-    handleDeprecatedParameters(mavenBuildLogger);
+    MavenJibLogger mavenJibLogger = new MavenJibLogger(getLog());
+    handleDeprecatedParameters(mavenJibLogger);
     SystemPropertyValidator.checkHttpTimeoutProperty(MojoExecutionException::new);
 
     if (!new DockerClient().isDockerInstalled()) {
@@ -70,20 +70,20 @@ public class BuildDockerMojo extends JibPluginConfiguration {
 
     // Parses 'from' and 'to' into image reference.
     MavenProjectProperties mavenProjectProperties =
-        MavenProjectProperties.getForProject(getProject(), mavenBuildLogger, getExtraDirectory());
+        MavenProjectProperties.getForProject(getProject(), mavenJibLogger, getExtraDirectory());
     ImageReference baseImage = parseImageReference(getBaseImage(), "from");
     ImageReference targetImage =
-        mavenProjectProperties.getGeneratedTargetDockerTag(getTargetImage(), mavenBuildLogger);
+        mavenProjectProperties.getGeneratedTargetDockerTag(getTargetImage(), mavenJibLogger);
 
     // Checks Maven settings for registry credentials.
     if (Boolean.getBoolean("sendCredentialsOverHttp")) {
-      mavenBuildLogger.warn(
+      mavenJibLogger.warn(
           "Authentication over HTTP is enabled. It is strongly recommended that you do not enable "
               + "this on a public network!");
     }
     MavenSettingsServerCredentials mavenSettingsServerCredentials =
         new MavenSettingsServerCredentials(
-            Preconditions.checkNotNull(session).getSettings(), settingsDecrypter, mavenBuildLogger);
+            Preconditions.checkNotNull(session).getSettings(), settingsDecrypter, mavenJibLogger);
     Authorization fromAuthorization = getBaseImageAuth();
     RegistryCredentials knownBaseRegistryCredentials =
         fromAuthorization != null
@@ -111,13 +111,13 @@ public class BuildDockerMojo extends JibPluginConfiguration {
             .setEnvironment(getEnvironment())
             .setExposedPorts(ExposedPortsParser.parse(getExposedPorts()));
     if (getUseCurrentTimestamp()) {
-      mavenBuildLogger.warn(
+      mavenJibLogger.warn(
           "Setting image creation time to current time; your image may not be reproducible.");
       containerConfigurationBuilder.setCreationTime(Instant.now());
     }
 
     BuildConfiguration.Builder buildConfigurationBuilder =
-        BuildConfiguration.builder(mavenBuildLogger)
+        BuildConfiguration.builder(mavenJibLogger)
             .setBaseImageConfiguration(baseImageConfiguration)
             .setTargetImageConfiguration(targetImageConfiguration)
             .setContainerConfiguration(containerConfigurationBuilder.build())
@@ -135,7 +135,7 @@ public class BuildDockerMojo extends JibPluginConfiguration {
     BuildConfiguration buildConfiguration = buildConfigurationBuilder.build();
 
     // TODO: Instead of disabling logging, have authentication credentials be provided
-    MavenBuildLogger.disableHttpLogging();
+    MavenJibLogger.disableHttpLogging();
 
     RegistryClient.setUserAgentSuffix(USER_AGENT_SUFFIX);